# proyecto-dws-grupo2/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: TitanGym API
  description: Documentación oficial de la API REST del gimnasio TitanGym
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/classes/{id}:
    get:
      tags:
      - Clases
      summary: Obtener una clase por su ID
      description: Busca una clase específica utilizando su identificador único.
      operationId: getClassById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: No existe ninguna clase con ese ID
          content: {}
        "200":
          description: Clase encontrada exitosamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
    put:
      tags:
      - Clases
      summary: Actualizar una clase existente
      description: Modifica los datos completos de una clase ya registrada.
      operationId: updateClass
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassRequestDTO"
        required: true
      responses:
        "404":
          description: Clase no encontrada para actualizar
          content: {}
        "200":
          description: Clase actualizada correctamente
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
        "400":
          description: Datos de actualización inválidos
    delete:
      tags:
      - Clases
      summary: Eliminar una clase
      description: Borra permanentemente una clase del sistema.
      operationId: deleteClass
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Clase no encontrada para eliminar
          content: {}
        "200":
          description: Clase eliminada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
...
  /api/v1/reviews/:
    get:
      tags:
      - Reseñas
      summary: Obtener todas las reseñas
      description: Devuelve una página con todas las valoraciones de los clientes.
      operationId: getReviews
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: Lista de reseñas recuperada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelReviewResponseDTO"
    post:
      tags:
      - Reseñas
      summary: Publicar una nueva reseña
      description: Crea un comentario y una puntuación asociada a una clase y un usuario.
      operationId: createReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Reseña publicada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponseDTO"
        "400":
          description: Datos de la reseña inválidos
  /api/v1/classes:
    get:
      tags:
      - Clases
      summary: Obtener todas las clases
      description: Devuelve una página con todas las clases o actividades disponibles
        en el gimnasio.
      operationId: getAllClasses
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "404":
          description: Not Found
        "200":
          description: Lista de clases recuperada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelClassResponseDTO"
    post:
      tags:
      - Clases
      summary: Crear una nueva clase
      description: Añade una nueva clase o actividad al catálogo del gimnasio.
      operationId: createClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassRequestDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "201":
          description: Clase creada con éxito
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ClassResponseDTO"
        "400":
          description: Los datos proporcionados para la clase son inválidos
...
  /api/v1/classes/{id}/reviews:
    get:
      tags:
      - Clases
      summary: Obtener reseñas de una clase
      description: Recupera todas las valoraciones y comentarios asociados a una clase
        concreta.
      operationId: getClassReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Clase no encontrada
          content: {}
        "200":
          description: Lista de reseñas recuperada con éxito
          content:
            '*/*':
              schema:
                type: object
  /api/v1/classes/{id}/attendees:
    get:
      tags:
      - Clases
      summary: Obtener asistentes de una clase
      description: Recupera la lista de todos los usuarios inscritos en una clase
        específica.
      operationId: getClassAttendees
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Clase no encontrada
          content: {}
        "200":
          description: Lista de asistentes recuperada con éxito
          content:
            '*/*':
              schema:
                type: object
components:
  schemas:
...
    ImageResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        url:
          type: string
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        profileImage:
          $ref: "#/components/schemas/ImageResponseDTO"
...
    ReviewRequestDTO:
      type: object
      properties:
        about:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        userId:
          type: integer
          format: int64
        classEntityId:
          type: integer
          format: int64
    ReviewResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        about:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        user:
          $ref: "#/components/schemas/UserResponseDTO"
    ClassRequestDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        schedule:
          type: string
    ClassResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        schedule:
          type: string
        image:
          $ref: "#/components/schemas/ImageResponseDTO"
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
...
    PagedModelReviewResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReviewResponseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
    PagedModelClassResponseDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/ClassResponseDTO"
        page:
          $ref: "#/components/schemas/PageMetadata"
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT